Update Vortice.DirectX NuGet packages

+++ DirectX12GameEngine.Graphics/GraphicsPresenter.cs
@@ -46,7 +46,7 @@ namespace DirectX12GameEngine.Graphics
         {
             return Texture.New2D(GraphicsDevice,
                 PresentationParameters.BackBufferWidth, PresentationParameters.BackBufferHeight, PresentationParameters.DepthStencilFormat,
-                TextureFlags.DepthStencil, 1, PresentationParameters.Stereo ? 2 : 1);
+                TextureFlags.DepthStencil, 1, PresentationParameters.Stereo ? (short)2 : (short)1);
         }
 
         protected abstract void ResizeBackBuffer(int width, int height);
